+++ framework/db-local/src/main/java/is/codion/framework/db/local/logger/MethodLogger.java
@@ -49,10 +49,6 @@ public interface MethodLogger {
 		return new DefaultMethodLogger(maxSize, formatter);
 	}
 
-	static MethodLogger noOpLogger() {
-		return DefaultMethodLogger.NO_OP;
-	}
-
 	interface Loggable {
 
 		void methodLogger(@Nullable MethodLogger methodLogger);